// client/src/component/Common/Layouts/Box/Box.types.ts
import React from 'react';
import { CSSProperties } from 'react';

export interface BoxProps extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
  bold?: boolean;
  // css
  display?: CSSProperties['display'];
  flex?: CSSProperties['flex'];
  alignItems?: CSSProperties['alignItems'];
  justifyContent?: CSSProperties['justifyContent'];
  flexDirection?: CSSProperties['flexDirection'];
  flexWrap?: CSSProperties['flexWrap'];
  gap?: CSSProperties['gap'];
  position?: CSSProperties['position'];
  left?: CSSProperties['left'];
  right?: CSSProperties['right'];
  top?: CSSProperties['top'];
  bottom?: CSSProperties['bottom'];
  m?: CSSProperties['margin'];
  mh?: CSSProperties['marginLeft'];
  mv?: CSSProperties['marginTop'];
  mt?: CSSProperties['marginTop'];
  mb?: CSSProperties['marginBottom'];
  ml?: CSSProperties['marginLeft'];
  mr?: CSSProperties['marginRight'];
  p?: CSSProperties['padding'];
  ph?: CSSProperties['paddingLeft'];
  pv?: CSSProperties['paddingTop'];
  pt?: CSSProperties['paddingTop'];
  pb?: CSSProperties['paddingBottom'];
  pl?: CSSProperties['paddingLeft'];
  pr?: CSSProperties['paddingRight'];
  width?: CSSProperties['width'];
  height?: CSSProperties['height'];
  minWidth?: CSSProperties['minWidth'];
  maxWidth?: CSSProperties['maxWidth'];
  minHeight?: CSSProperties['minHeight'];
  maxHeight?: CSSProperties['maxHeight'];
  fontFamily?: CSSProperties['fontFamily'];
  fontSize?: CSSProperties['fontSize'];
  fontWeight?: CSSProperties['fontWeight'];
  lineHeight?: CSSProperties['lineHeight'];
  textAlign?: CSSProperties['textAlign'];
  color?: CSSProperties['color'];
  backgroundColor?: CSSProperties['backgroundColor'];
  background?: CSSProperties['background'];
  backgroundImage?: CSSProperties['backgroundImage'];
  backgroundSize?: CSSProperties['backgroundSize'];
  backgroundPosition?: CSSProperties['backgroundPosition'];
  backgroundRepeat?: CSSProperties['backgroundRepeat'];
  border?: CSSProperties['border'];
  borderTop?: CSSProperties['borderTop'];
  borderBottom?: CSSProperties['borderBottom'];
  borderLeft?: CSSProperties['borderLeft'];
  borderRight?: CSSProperties['borderRight'];
  borderRadius?: CSSProperties['borderRadius'];
  overflow?: CSSProperties['overflow'];
  opacity?: CSSProperties['opacity'];
  zIndex?: CSSProperties['zIndex'];
  transform?: CSSProperties['transform'];
}